Interpolation method changed. Now the counts are calculated from the concentration due to an issue how the data stream is sent to the tanks
We calculate the raw counts from the CO concentration and the calibration coefficients.
The AL52CO_counts variable can not be used because it does not necessarily match the
concentration. The data stream that is sent by the labview script is sent several times
a second while only *one* value in the stream is updated at a time. For example see:

$ cat AL52CO01_20140126_060103_B828.cs
...
$AL52CO01,66,1390716327,0,0,B828,77.721321,8821.000000, ...
$AL52CO01,66,1390716327,0,0,B828,77.721321,8475.000000, ...
$AL52CO01,66,1390716328,0,0,B828,77.721321,8475.000000, ...
$AL52CO01,66,1390716328,0,0,B828,81.419006,8475.000000, ...
$AL52CO01,66,1390716328,0,0,B828,81.419006,8660.000000, ...
$AL52CO01,66,1390716329,0,0,B828,81.419006,8660.000000, ...
...

The concentration changes in the forth line to 81.42 but the counts value is only update
the subsequent line.